+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# NEXT 
+
+## New Features
+
+- Improved proof point validation API
+
+## Bug Fixes
+
+None
+
+## Breaking Changes
+
+- `proofPoint.validate` and `proofPoint.validateByHash` now return the new type `ProofPointVlaidationResult` giving more information about the status of the proof point.
+
 # 2.0.0
 
 ## New Features